feat: 添加无线通信模块及相关初始化函数

+++ code/algo/img/image.c
@@ -187,7 +187,7 @@ void computeCentroid(uint16* labeledImage,
     }
 }
 
-void drawRedPlus(uint16* image,
+void drawRedPlus(uint8* image,
                  uint16 width,
                  uint16 height,
                  uint16 cx,
